pyproject.toml
[project]
name = "itwhom"
version = "0.1.0"
description = "Add your description here"
authors = [
  { name = "Taylor Monacelli", email = "taylormonacelli@gmail.com" }
]
dependencies = [
  "jinja2>=3.1.3",
  "python-frontmatter",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
itwhom = "itwhom:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
  "black>=24.2.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/itwhom"]
